Historybound

Elo | 15.11 +- 7.37 (95%)
SPRT | 8.0+0.08s Threads=1 Hash=8MB
LLR | 2.94 (-2.94, 2.94) [0.00, 5.00]
Games | N: 3222 W: 905 L: 765 D: 1552
Penta | [43, 341, 729, 429, 69]
https://openbench.tstawowy.pl/test/35/

bench: 3067366

+++ src/search.rs
@@ -358,7 +358,7 @@ pub fn pvs<NODE: NodeType>(board_position: &BoardPosition, search_state: &mut Se
 
             // Often reduce less for good-history moves
             //search_state
-            reduction -= search_state.get_quiet_history(board_position.side, mv) / 8;
+            reduction -= search_state.get_quiet_history(board_position.side, mv) as i32 / 8;
 
             let reduction = (reduction / 1024).clamp(0, (depth - 1) as i32) as usize;
